Vote on first 3 days vs last 3 days of month
This is a pragmatic decision designed to make scheduling voting periods
easier. It is difficult to create recurring events in calendar
applications for "the last 3 days of the month", but easy to do so for
"the first 3 days of the month". This change also gives contributors
until the calendar end of the month to get their contribution requests
submitted, which is also likely to be easier to remember and deal with
in general.

This commit also updates the date of the first voting period to be
November 6th, 2017 (instead of November 1st). This is simply to allow
more time for communication and preparation.