Move phase-zero/README.adoc => phase-zero.adoc
This effectively reverts cb9e11ce6cf1524163909366480b934ecd2202a0, in
which I moved dao/phase-zero.adoc => dao/phase-zero/README.adoc. I liked
this approach in principle because it allowed for cleaner, suffix-free
GitHub links, e.g. https://github.com/bisq-network/docs/dao/phase-zero,
in which the README.adoc would automatically render for the user, and in
which all the files in that directory (including) images would be listed
at the top, such that readers could see all recent changes to the doc.
This approach kept the doc "self contained" within a single (phase-zero)
directory, and there were other benefits as well. However, after living
with this for a few days, I've changed my mind, mainly because it's not
an obvious approach for other authors, and will probably be an uphill
battle to try to get everyone to write docs this way; it might have nice
effects, but explaining and trying to enforce the approach just wouldn't
worth it. So this change reverts it, and just goes with the simple,
obvious approach of writing these docs in named files, e.g.
phase-zero.adoc, with an optional directory parallel to the file, e.g.
'phase-zero/' when the doc has images or other assets to reference.