Remove 'rebase and redistribute BSQ' monthly action
We (@ManfredKarrer and I) decided not to continue with this process as
it was too cumbersome / time consuming and did not add enough value /
protection to justify that effort. We are now simply letting the total
BSQ supply grow in the BSQ stake and vote tracking spreadsheet, and we
will do a final rebasing and redistribution to 2.5 MM BSQ when going
live on mainnet.