Streamline token listing process
Problem: We get a lot of token listing pull requests, and they consist
of any number of commits with mixed comment styles and inconsistent
author metadata.  This pollutes the commit history and prevents us from
being able to cleanly revert a token listing with a single commit.

Solution: This commit adds a step to list-token.adoc that instructs the
contributor to form a single, atomic commit containing their changes and
to use proper Git author metadata when doing so. It also removes some
information from the pull request template that is now redundant given
the instruction to include both the token name and ticker symbol in the
commit title (which will by default become the pull request title now
that they will be submitting a single commit).