Eliminate 'src' dir
I put everything in src to make it possible to do Gradle continuous
builds, but in the end this isn't worth it. The actual asciidoctor build
is too slow, and even with LiveReload in the mix, the turnaround time
wasn't sufficient. In the end it's better to go with instantaneous rendering
in the browser with the Asciidoctorjs Chrome plugin, and do full `gradle
asciidoctor` runs on an as-needed basis.

As a side benefit, the existing 'dao' and 'exchange' docs "just work"
once again, without needing to be moved. The idea for now is to keep
everything as flat as possible, right in the root directory, but no harm
keeping these docs where they are for a bit while writing everything
else.