Produce a 'new contributor checklist'
This is all the stuff we very often point new contributors to over the
first days and weeks of their involvement with Bisq. The motivation of
writing this doc is to have a single link we can provide to future new
contributors, such that they can ramp up autonomously and at their own
pace, making everyone involved more productive.

There's a lot we can do to make this doc more friendly, less dense, etc.
It's a first cut; a "something is better than nothing" best effort to
get the ball rolling.