Add instructions for generating a new BSQ address
When the Phase Zero doc was originally written, the audience was _past
contributors:_ those who had received a portion of the initial
distribution of BSQ, and they had been given special instructions how to
generate a BSQ address.

This change updates the Phase Zero doc to include the same instructions
for the benefit of new contributors submitting their first compensation
requests.