Use --appName to avoid collision with default Bisq app
Previously, the instructions would spin up an API instance that would
attempt to use the default app name and app data directory, meaning that
it would run into filesystem locks with an existing local Bisq app if
one is running. We should assume that users interested in using the Bisq
HTTP API are already running Bisq locally, and in order to avoid this
kind of collision, and also to just generally play it safe, we should
instruct the user to spin up their instance in a throwaway data
directory.

I've used the name 'bisqmon' here as in, a "Bisq monitor bot".

This commit does *not* update the `docker run` invocation in the same
way, but it should be updated to match.