Verify the API is up with 'version' endpoint
Previously, the user was instructed to verify the API is up and running
with the 'offers' endpoint and was then instructed in the next section
to invoke the same endpoint.

This change makes it easy to see that the API is up and running by
returning a much shorter and easier to read response, and has the added
side benefit of informing the user which version of Bisq they're
running.

The user then moves on in the next step to explore the 'offers'
endpoint, which they're now more ready to do.