Add 'improve this doc' section
With link to document source in GitHub. Also, consolidate 'Get help' and
'Stay in touch' sections in order to cut down on these extra sections at
the bottom of the document, and because the line between getting help
and staying in touch isn't so clear cut (Slack, for example is really
both).