Flesh out 'User Docs' section of index page
As explained in the new admonition at the top of the page, many of these
docs have not yet been written, and therefore have no links. The idea is
to publish this list of the docs we have _and_ the docs we want to have,
such that users can get a sense of everything we want them to know about
Bisq, even though it's not all ready to go yet, and such that
contributors can get a sense of where they can add value by writing
these incomplete docs.

In subsequent commits, the same kind of 'fleshing out' will be done for
the 'Contributor Docs' section of the index page.