Continue fleshing out 'Glossary'
This was the result of a lot of experimentation with formats for the
Glossary in the context of building out the Payment Methods doc. In the
end, the best is to structure it just like any other doc, i.e. no
special `term:: definition` syntax, as it breaks expectations like
ability to anchor link to the terms.

It was also an excercise in developing the tersest and most digestible
format possible. What's coming together here is inspired by @evoskuil's
work at https://github.com/libbitcoin/libbitcoin/wiki/Glossary.

In any case it remains very much a work in progress, but one worth
committing at this point.