Simplify date metadata in phase zero doc
A quirk in Asciidoctor forces you to specify a version in order to also
write out a date in long form. This doc has no versioning other than its
git history, so it's misleading to see the `v1.0`. This change replaces
it with a simpler `2017-10-19`, which Asciidoctor presents correctly and
without any requirement for a version string.