Customize [example] style in glossary docinfo
Previously, [example] blocks rendered with a pale yellow that generally
looks good, but is a bit too eye-catching for the purposes of this
document. Asciidoctor's [sidebar] block style was closer to the mark,
but the semantics of [example] are still desirable, as that's exactly
what these blocks represent: example usages. So this change customizes
[example] blocks to use the same background color as sidebar blocks, and
only for this doc through the private docinfo mechanism [1]. It also
customizes the coloration of `code` text, as the default coloration
there is basically the same as the background, making the two wash out.

[1]: https://asciidoctor.org/docs/user-manual/#docinfo-file